/ppbox/branches/1.2.0 /src/ 1、修正 avcodec/Assembler.h 工厂 map 关键字类型错误，造成创建的 Assembler 混乱 2、修正 capture/CaptureSource 支持多线程修改的一些问题 emailto: box-dev
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@6959 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ Assembler.h
@@ -16,7 +16,7 @@ namespace ppbox
         class Assembler
             : public ppbox::common::ClassFactory<
                 Assembler, 
-                boost::uint32_t, 
+                boost::uint64_t, 
                 Assembler *()
             >
         {
